requirements.txt
asgiref==3.7.2
Django==4.2.6
django-sesame==3.2.1
psycopg==3.1.12
psycopg-binary==3.1.12
sqlparse==0.4.4
typing_extensions==4.8.0
